requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==2.4.0
channels-redis==2.4.2
constantly==15.1.0
cryptography==3.4.7
daphne==2.5.0
Django==3.0.7
django-filter==2.2.0
djangorestframework==3.11.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
Markdown==3.2.1
msgpack==0.6.2
mysqlclient==1.4.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
zope.interface==5.4.0
